requirements.txt
python-telegram-bot
requests
beautifulsoup4
lxml
feedparser
python-dotenv
google-generativeai
schedule
pytest
openai
